Fix performance issue (#17)
The problem is that as the evaluation went on the time required for evaluating on test. After profiles it showed that torch_macs was the culprit, meaning that torch_macs was being executed on test. That however, didn't explain why the time increased in each iteration. After some meditating I think that the implementation of the mode made that if a new mode was created each time prehook functions would stack on each model firing causing an ever increasing need for resources. This commit should fix that. As a TODO it would be that an extra cost would still apply in test even if torch_macs isn't used, but, at least, it won't be an increasing cost.